[DPP-25] Changed all Bang references to Http
- Issue:
The initial data source implementation was called BangDataSource,
because it was used only to retreive HTML pages from Banggood. However
the naming was misleading because the implementation is able to make
HTTP calls to any url.

- Outcome:
Changed all reference to Bang in Http so that the naming scheme is now
semantically correct.