[DPP-92] Modified response page closing in snapshot building
Issue
If the source URL is non-existing there is no response body, so the body
can't be closed. The previous codebase deferred the body closing
function regardless of the existence of the body.

Code changes
Added deferring of the close response body function in the else branch
of error management of the http fetch function.