[DPP-107] Returning a string instead of a number
The issue was caused by a function in the Ebay service that was
returning a string when an item is active and an integer when the item
is inactive. This had a cascading effect on the main app as that field
had String as type but sometimes was receving an Int. When trying to
deserialise the object the runtime failed and marked the item as unkown